[wip] add cheerio library
