+++ package.json
@@ -5,6 +5,7 @@
     "@rails/webpacker": "^3.3.1",
     "babel-preset-react": "^6.24.1",
     "bootstrap-sass": "^3.3.7",
+    "cheerio": "^1.0.0-rc.2",
     "es6-shim": "^0.35.3",
     "exports-loader": "^0.6.4",
     "font-awesome": "^4.7.0",
